Fix bug in fitsData::readFile, mask is unsigned int
* Fix bug when reading file with mask where a ++cntr
  should have been a cntr++, which was causing a memory leak.
* Add unit test for SPIRE-like maps
* Explicitly read mask as unsigned int.  This may cause
  portability issues, since FITS does't properly support
  unsigned values.
* Add missing #include<limits> to a bunch of the example code
* Change some types to match cfitsio in fitsData and fitsDataDouble --
  for example, number of axes should be a long rather than
  unsigned int -- yes, that is utterly ridiculous, but it's what cfitsio
  uses.

+++ include/fitsDataDouble.h
@@ -32,8 +32,8 @@ class fitsDataDouble {
   unsigned int* binval; //!< Number of elements in each bin, row major array
 
   /*! \brief Reads data from single file*/
-  bool readFile(const std::string&,unsigned int&,double*&, 
-		int*&,bool=false);
+  bool readFile(const std::string&, long&, double*&, 
+		unsigned int*&, bool=false);
 
 
  public:
